Summary: Document arable/forest land, money fungibility, and building overhaul
Body:
- Add /game_balance_analyzer/build/ to .gitignore.
- Update AGENTS.md with extensive documentation on new game mechanics:
  - Arable land and forest land as abstract resources limiting building placement, derived from manor level.
  - Money fungibility (gold and silver pence interchangeable) and auto-import of build materials.
  - Building overhaul: stage chains via `built_from`, class-based grouping, level-indexed production arrays, manor house ledger (10 tiers), stone removal, and `descriptions` internal-note field.
  - Analyzer-only config convention (`analyzer_` prefix) and validation rules.
- Update api/Build.md with the new `convert` action, error codes (`insufficient_arable_land`, `insufficient_forest_land`), auto-import behavior, and updated cost fields (removed stone, added beams/boards/iron/ironwork).

+++ server/Money.hpp
@@ -24,9 +24,10 @@ long gold_to_pence(double gold);
  *
  * @param gold - In/out: gold pieces (carries any silver overflow, discards the
  *               fractional pence remainder)
- * @param silver_pence - In/out: pence balance, normalized to 0..239
+ * @param silver_pence - In/out: pence balance (fractional-capable), normalized
+ *               to 0..239
  */
-void normalize(double& gold, int& silver_pence);
+void normalize(double& gold, double& silver_pence);
 
 /**
  * Converts any object-form money cost entry into a gold double, in place.
